+++ CHANGES.rst
@@ -22,6 +22,8 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
   `CompoundCriterion` adapter `items-with-advice-back-to-item-validation-states`
   to get items having finances advice that are return in item validation states.
   [gbastien]
+- Adapted code regarding removal of `MeetingConfig.useGroupsAsCategories`.
+  [gbastien]
 
 4.2b24 (2022-09-29)
 -------------------