+++ CHANGES.rst
@@ -5,8 +5,8 @@ Products.MeetingCommunes Changelog
 The Products.MeetingCommunes version must be the same as the Products.PloneMeeting version
 
 
-4.2b25 (unreleased)
--------------------
+4.2 (unreleased)
+----------------
 
 - Removed useless import of `get_cachekey_volatile` in `adapters.py`.
   [gbastien]